+ Cambiado el float de duración de todas las clases a int. Ahora duración = 1 significa semifusa y no se puede ir más abajo en la jerarquía. + Purgado el atributo Clave. De ahora en adelante es Tonalidad y solo la tiene Segmento. Clave no tiene sentido (Clave de fa, clave de sol,... es solo para aspecto gráfico). + Ahora la tabla es una clase aunque con posibilidades de cambiar de nombre. + transformNota de MidizatorAbc se hace bien, aunque falta hacerle pruebas... - Falta añadir que se quiten los accidentes(# y b) despues de que termine un compás.

+++ trunk/Muphic/Mu/source/include/Acorde.h
@@ -9,16 +9,16 @@ class Acorde : public Simbolo
 {
 	public:
 		Acorde();
-		Acorde(float d);
-		Acorde(float d, int n1, int n2);
-		Acorde(float d, int n1, int n2, int n3);
-		Acorde(float d, int n1, int n2, int n3, int n4);
+		Acorde(int d);
+		Acorde(int d, int n1, int n2);
+		Acorde(int d, int n1, int n2, int n3);
+		Acorde(int d, int n1, int n2, int n3, int n4);
 
 
 		Acorde(Nota* notas);
 		Acorde(Nota notas[], int length);
-		Acorde(float d, Nota* notas);
-		Acorde(float d, Nota notas[], int length);
+		Acorde(int d, Nota* notas);
+		Acorde(int d, Nota notas[], int length);
 
 		~Acorde();
